Flyway

+++ db/index.js
@@ -3,75 +3,58 @@ var path = require('path');
 var fs = require('fs');
 var async = require('async');
 var config = require("../config");
+var exec = require('child_process').exec;
 var db = mysql.createConnection({
 	host: config.ENV.database.host,
 	user: config.ENV.database.user,
 	password: config.ENV.database.password,
 	multipleStatements: true
 });
+const baseCmd = `flyway -user=${config.ENV.database.user} -password=${config.ENV.database.password} -url=jdbc:mysql://${config.ENV.database.host}/${config.ENV.database.database}`;
+const locations = `-locations=filesystem:${__dirname}/sql,filesystem:${__dirname}/`;
+
 var method = process.argv[2];
 var internals = {};
 
 internals._sqldir = __dirname + '/sql';
 
-internals._runfile = function _runfile(s, fname, callback, optional) {
-	var sql;
-
-	if (typeof fname == 'function') {
-		callback = fname;
-		fname = s;
-		sql = 'USE ' + config.ENV.database.database + '; \n'
-	} else {
-		sql = s;
-	}
-	var fpath = path.resolve(internals._sqldir, fname);
-	if (!fs.existsSync(fpath)) {
-		if (optional === true) {
-			return callback();
-		}
-
-		return callback(Error('sql file not found - ' + fpath));
-	}
-	if (!fs.lstatSync(fpath).isFile()) return callback(Error(fpath + ' is not a file'));
-	try {
-		sql += fs.readFileSync(fpath, 'utf8');
-	}
-	catch (e) {
-		return callback(Error('error reading ' + fpath + ' - ' + e.stack))
-	}
-	if (!sql.trim()) return callback(Error('empty file - ' + fpath));
-	console.log('Running file: ' + fpath);
-	db.query(sql, callback);
-};
-
-internals.schema = function schema() {
+internals._runfile = function _runfile(s, callback) {
 	console.log('db.schema');
-	var query = '';
 
-	query += 'DROP DATABASE IF EXISTS ' + config.ENV.database.database + '; \n';
-	query += 'CREATE DATABASE IF NOT EXISTS ' + config.ENV.database.database + ' DEFAULT CHARACTER SET utf8; \n';
-	query += 'USE ' + config.ENV.database.database + ';';
+	async.waterfall([
+		(done) => {
+			var query = '';
+			if (s == 'test') {
+				query += 'DROP DATABASE IF EXISTS ' + config.ENV.database.database + '; \n';
+				query += 'CREATE DATABASE IF NOT EXISTS ' + config.ENV.database.database + ' DEFAULT CHARACTER SET utf8; \n';
+				query += 'USE ' + config.ENV.database.database + ';';
+			} else {
+				query += 'CREATE DATABASE IF NOT EXISTS ' + config.ENV.database.database + ' DEFAULT CHARACTER SET utf8; \n';
+			}
+			db.query(query, err=> {
+				done(err)
+			});
+		},
+		(done)=> {
+			exec('echo $PATH', (e, stdout) => {
+				console.log("path: ", stdout);
+				if (e) return done(e);
+				exec(baseCmd + ' ' + locations + s + ` migrate`, function (err, stdout) {
+					console.log(stdout);
+					done(err);
+				});
+			});
+		}
+	], (err)=> {
+		callback(err);
+	})
 
-	db.query(query, function (e, r) {
-		if (e) throw e;
-		internals._runfile('schema.sql', function (e, r) {
-			if (e) throw e;
-			process.exit(0);
-		});
-	});
 };
 
-internals.base = function base() {
-	console.log('db.base');
-	internals._runfile('data-base.sql', function (e, r) {
-		if (e) throw e;
-		process.exit(0);
-	});
-};
 
 internals.test = function test() {
 	console.log('db.test');
-	internals._runfile('data-test.sql', function (e, r) {
+	internals._runfile('test', function (e, r) {
 		if (e) throw e;
 		process.exit(0);
 	});
@@ -79,7 +62,7 @@ internals.test = function test() {
 
 internals.local = function local() {
 	console.log('db.local');
-	internals._runfile('data-local.sql', function (e, r) {
+	internals._runfile('local', function (e, r) {
 		if (e) throw e;
 		process.exit(0);
 	});
@@ -87,26 +70,17 @@ internals.local = function local() {
 
 internals.dev = function dev() {
 	console.log('db.dev');
-	internals._runfile('data-dev.sql', function (e, r) {
+	internals._runfile('dev', function (e, r) {
 		if (e) throw e;
 		process.exit(0);
 	});
 };
 
-internals.truncate = function truncate() {
-	console.log('db.truncate');
-	var t, tables = config.db.tables;
-	var query = '';
-
-	query += 'USE ' + config.ENV.database.database + '; \n';
-	query += 'SET foreign_key_checks = 0; \n';
-	for (t in tables) {
-		query += "TRUNCATE `" + tables[t] + "`; \n";
-	}
-
-	query += 'SET foreign_key_checks = 1; \n';
-	db.query(query, function (e, r) {
-		if (e) throw e;
+internals.clean = function clean() {
+	console.log('db.clean');
+	exec(baseCmd + ' clean', function (err, stdout) {
+		console.log(stdout);
+		if (err) throw err;
 		process.exit(0);
 	});
 };
